server/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
google-cloud-texttospeech==2.34.0
google-cloud-storage==3.9.0
google-auth==2.42.0
gunicorn==25.1.0
python-dotenv==1.0.0
